fix: Break down most of liability.py
