+++ changelog_entry.yaml
@@ -2,3 +2,4 @@
   changes:
     fixed:
     - Folder distribution and formatting for income tax-related variables
+    - pays_scottish_income_tax now returns a Boolean value
