+++ policyengine_uk/tests/policy/integration/income.yaml
@@ -1,0 +1,34 @@
+- name: Earned income is taxed correctly
+  period: 2020
+  absolute_error_margin: 0
+  input:
+    employment_income: 100000
+  output:
+    allowances: 12500
+    earned_taxable_income: 87500
+    basic_rate_earned_income_tax: 7500
+    higher_rate_earned_income_tax: 20000
+    add_rate_earned_income_tax: 0
+    earned_income_tax: 27500
+    tax_band: "HIGHER"
+    taxed_income: 87500
+- name: Savings slice of basic rate band
+  period: 2020
+  absolute_error_margin: 0
+  input:
+    employment_income: 40000
+    savings_interest_income: 15000
+  output:
+    basic_rate_savings_income_pre_starter: 10000
+    savings_starter_rate_income: 0
+    basic_rate_savings_income: 10000
+    savings_allowance: 500
+    higher_rate_savings_income: 4500
+    taxed_savings_income: 14500
+- name: Dividends taxed at the marginal rates
+  period: 2020
+  absolute_error_margin: 0
+  input:
+    dividend_income: 50000
+  output:
+    dividend_income_tax: 6225
